requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-crispy-forms==1.14.0
django-embed-video==1.4.8
django-import-export==3.0.2
django-jazzmin==2.6.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.2
tablib==3.3.0
tzdata==2022.2
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
